Refactor / remove Affinity Group UI
