+++ app/pages/project/affinity/AntiAffinityGroupPage.tsx
@@ -10,6 +10,7 @@ import { createColumnHelper } from '@tanstack/react-table'
 import type { LoaderFunctionArgs } from 'react-router'
 
 import {
+  apiq,
   getListQFn,
   queryClient,
   usePrefetchedQuery,
@@ -20,14 +21,12 @@ import {
   getAntiAffinityGroupSelector,
   useAntiAffinityGroupSelector,
 } from '~/hooks/use-params'
-import {
-  AffinityGroupEmptyState,
-  antiAffinityGroupView,
-  GroupPage,
-} from '~/pages/project/affinity/utils'
+import { AffinityGroupEmptyState, AffinityPageHeader } from '~/pages/project/affinity/utils'
 import { makeLinkCell } from '~/table/cells/LinkCell'
 import { Columns } from '~/table/columns/common'
 import { useQueryTable } from '~/table/QueryTable'
+import { Badge } from '~/ui/lib/Badge'
+import { PropertiesTable } from '~/ui/lib/PropertiesTable'
 import { pb } from '~/util/path-builder'
 import type * as PP from '~/util/path-params'
 
@@ -38,10 +37,12 @@ export const handle = makeCrumb(
 
 const colHelper = createColumnHelper<AntiAffinityGroupMember>()
 
-const memberList = (query: PP.AntiAffinityGroup) =>
+const antiAffinityGroupView = ({ project, antiAffinityGroup }: PP.AntiAffinityGroup) =>
+  apiq('antiAffinityGroupView', { path: { antiAffinityGroup }, query: { project } })
+const memberList = ({ project, antiAffinityGroup }: PP.AntiAffinityGroup) =>
   getListQFn('antiAffinityGroupMemberList', {
-    path: { antiAffinityGroup: query.antiAffinityGroup },
-    query: { project: query.project },
+    path: { antiAffinityGroup: antiAffinityGroup },
+    query: { project: project },
   })
 
 export async function clientLoader({ params }: LoaderFunctionArgs) {
@@ -59,6 +60,7 @@ export default function AntiAffinityPage() {
   const { data: group } = usePrefetchedQuery(
     antiAffinityGroupView(antiAffinityGroupSelector)
   )
+  const { id, name, description, policy, timeCreated } = group
   const { data: members } = usePrefetchedQuery(
     memberList(antiAffinityGroupSelector).optionsFn()
   )
@@ -74,11 +76,26 @@ export default function AntiAffinityPage() {
   const { table } = useQueryTable({
     query: memberList(antiAffinityGroupSelector),
     columns,
-    emptyState: <AffinityGroupEmptyState type="anti-affinity" />,
+    emptyState: <AffinityGroupEmptyState />,
     getId: (member: AntiAffinityGroupMember) => member.value.id,
   })
 
   return (
-    <GroupPage type="anti-affinity" group={group} table={table} members={membersCount} />
+    <>
+      <AffinityPageHeader name={name} />
+      <PropertiesTable columns={2} className="-mt-8 mb-8">
+        <PropertiesTable.Row label="type">
+          <Badge>anti-affinity</Badge>
+        </PropertiesTable.Row>
+        <PropertiesTable.DescriptionRow description={description} />
+        <PropertiesTable.Row label="policy">
+          <Badge color="neutral">{policy}</Badge>
+        </PropertiesTable.Row>
+        <PropertiesTable.DateRow date={timeCreated} label="Created" />
+        <PropertiesTable.Row label="Members">{membersCount}</PropertiesTable.Row>
+        <PropertiesTable.IdRow id={id} />
+      </PropertiesTable>
+      {table}
+    </>
   )
 }
