+++ app/routes.tsx
@@ -488,38 +488,14 @@ export const routes = createRoutesFromElements(
           <Route
             path="affinity"
             lazy={() => import('./pages/project/affinity/AffinityPage.tsx').then(convert)}
-          >
-            <Route index element={<Navigate to="anti-affinity-groups" replace />} />
-            <Route
-              path="anti-affinity-groups"
-              lazy={() =>
-                import('./pages/project/affinity/AntiAffinityGroupsTab.tsx').then(convert)
-              }
-            />
-            <Route
-              path="affinity-groups"
-              lazy={() =>
-                import('./pages/project/affinity/AffinityGroupsTab.tsx').then(convert)
-              }
-            />
-          </Route>
+          />
           <Route path="affinity" handle={{ crumb: 'Affinity' }}>
-            <Route path="anti-affinity-groups" handle={{ crumb: 'Anti-affinity groups' }}>
-              <Route
-                path=":antiAffinityGroup"
-                lazy={() =>
-                  import('./pages/project/affinity/AntiAffinityGroupPage.tsx').then(convert)
-                }
-              />
-            </Route>
-            <Route path="affinity-groups" handle={{ crumb: 'Affinity groups' }}>
-              <Route
-                path=":affinityGroup"
-                lazy={() =>
-                  import('./pages/project/affinity/AffinityGroupPage.tsx').then(convert)
-                }
-              />
-            </Route>
+            <Route
+              path=":antiAffinityGroup"
+              lazy={() =>
+                import('./pages/project/affinity/AntiAffinityGroupPage.tsx').then(convert)
+              }
+            />
           </Route>
         </Route>
       </Route>
